update 45 13/05/2022 ARREGLOS EN LOG Y SEGMENTACION TEMATICA

+++ 1/lib/lib_system.php
@@ -5,11 +5,11 @@
 ** Funcion que genera logs de log-in
 */
 
-function logs($var){
+function logs($var,$ip_add){
       
       $fileName = "logs/$var.log.txt"; 
       $date = date("d/m/Y H:i:s");
-      $message = 'Ultimo ingreso: '.$date;
+      $message = 'Ultimo ingreso: '.$date. '- Desde la IP: '.$ip_add.'';
        
   if (file_exists($fileName)){
   
@@ -28,6 +28,136 @@ function logs($var){
 }
 
 
+
+    //función que escribe la IP del cliente en un archivo de texto    
+    function write_visita($var){
+
+        //Indicar ruta de archivo válida
+        $archivo = "logs/$var.log.txt";
+
+        //Si que quiere ignorar la propia IP escribirla aquí, esto se podría automatizar
+        $ip = "mi.ip.";
+        $new_ip = get_client_ip();
+
+        if ($new_ip !== $ip){
+            $now = new DateTime();
+
+       //Distinguir el tipo de petición, 
+       // tiene importancia en mi contexto pero no es obligatorio
+
+        if (!$_GET) {
+            $datos = "*POST: ".$_POST;
+
+        } 
+        else
+        {
+            //Saber a qué URL se accede
+            $peticion = explode('/', $_GET['PATH_INFO']);
+            $datos = str_pad($peticion[0],10).' '.$peticion[1];   
+        }
+        $txt =  str_pad($new_ip,25). " ".
+                str_pad($now->format('d-m-Y H:i:s'),25)." ".
+                str_pad(ip_info($new_ip, "Country"),25)." ".json_encode($datos);
+
+        $myfile = file_put_contents($archivo, $txt.PHP_EOL , FILE_APPEND);
+        }
+    }
+
+
+function get_client_ip() {
+        
+        $ipaddress = '';
+        if(getenv('HTTP_CLIENT_IP')){
+            $ipaddress = getenv('HTTP_CLIENT_IP');
+        }
+        else if(getenv('HTTP_X_FORWARDED_FOR')){
+            $ipaddress = getenv('HTTP_X_FORWARDED_FOR');
+        }
+        else if(getenv('HTTP_X_FORWARDED')){
+            $ipaddress = getenv('HTTP_X_FORWARDED');
+        }
+        else if(getenv('HTTP_FORWARDED_FOR')){
+            $ipaddress = getenv('HTTP_FORWARDED_FOR');
+        }
+        else if(getenv('HTTP_FORWARDED')){
+           $ipaddress = getenv('HTTP_FORWARDED');
+        }
+        else if(getenv('REMOTE_ADDR')){
+            $ipaddress = getenv('REMOTE_ADDR');
+        }
+        else{
+            $ipaddress = 'UNKNOWN';
+        }
+        
+        return $ipaddress;
+    }
+
+
+function ip_info($ip = NULL, $purpose = "location", $deep_detect = TRUE) {
+        $output = NULL;
+        if (filter_var($ip, FILTER_VALIDATE_IP) === FALSE) {
+            $ip = $_SERVER["REMOTE_ADDR"];
+            if ($deep_detect) {
+                if (filter_var(@$_SERVER['HTTP_X_FORWARDED_FOR'], FILTER_VALIDATE_IP))
+                    $ip = $_SERVER['HTTP_X_FORWARDED_FOR'];
+                if (filter_var(@$_SERVER['HTTP_CLIENT_IP'], FILTER_VALIDATE_IP))
+                    $ip = $_SERVER['HTTP_CLIENT_IP'];
+            }
+        }
+        $purpose    = str_replace(array("name", "\n", "\t", " ", "-", "_"), NULL, strtolower(trim($purpose)));
+        $support    = array("country", "countrycode", "state", "region", "city", "location", "address");
+        $continents = array(
+            "AF" => "Africa",
+            "AN" => "Antarctica",
+            "AS" => "Asia",
+            "EU" => "Europe",
+            "OC" => "Australia (Oceania)",
+            "NA" => "North America",
+            "SA" => "South America"
+        );
+        if (filter_var($ip, FILTER_VALIDATE_IP) && in_array($purpose, $support)) {
+            $ipdat = @json_decode(file_get_contents("http://www.geoplugin.net/json.gp?ip=" . $ip));
+            if (@strlen(trim($ipdat->geoplugin_countryCode)) == 2) {
+                switch ($purpose) {
+                    case "location":
+                        $output = array(
+                            "city"           => @$ipdat->geoplugin_city,
+                            "state"          => @$ipdat->geoplugin_regionName,
+                            "country"        => @$ipdat->geoplugin_countryName,
+                            "country_code"   => @$ipdat->geoplugin_countryCode,
+                            "continent"      => @$continents[strtoupper($ipdat->geoplugin_continentCode)],
+                            "continent_code" => @$ipdat->geoplugin_continentCode
+                        );
+                        break;
+                    case "address":
+                        $address = array($ipdat->geoplugin_countryName);
+                        if (@strlen($ipdat->geoplugin_regionName) >= 1)
+                            $address[] = $ipdat->geoplugin_regionName;
+                        if (@strlen($ipdat->geoplugin_city) >= 1)
+                            $address[] = $ipdat->geoplugin_city;
+                        $output = implode(", ", array_reverse($address));
+                        break;
+                    case "city":
+                        $output = @$ipdat->geoplugin_city;
+                        break;
+                    case "state":
+                        $output = @$ipdat->geoplugin_regionName;
+                        break;
+                    case "region":
+                        $output = @$ipdat->geoplugin_regionName;
+                        break;
+                    case "country":
+                        $output = @$ipdat->geoplugin_countryName;
+                        break;
+                    case "countrycode":
+                        $output = @$ipdat->geoplugin_countryCode;
+                        break;
+                }
+            }
+        }
+        return $output;
+    }
+    
 ///////////////////////////// SCRIPT PARA REALIZAR BACKUP /////////////////////////////////
 /*
 ** funcion para realizar backup de directorio
@@ -35,9 +165,10 @@ function logs($var){
 function backup(){
 
 	 $message = shell_exec("../../backup.sh");
-         echo '<div class="alert alert-success" role="alert">';
-	 echo '<h1 class="panel-title text-left" contenteditable="true"><img src="../../icons/actions/dialog-ok-apply.png"  class="img-reponsive img-rounded"><strong> '.print_r($message).'</strong></h1>';
-         echo "</div>";
+         echo '<div class="alert alert-success" role="alert">
+                <h1 class="panel-title text-left" contenteditable="true">
+                    <img src="../../icons/actions/dialog-ok-apply.png"  class="img-reponsive img-rounded"><strong> '.print_r($message).'</strong></h1>
+                </div>';
          
 
 }
@@ -46,19 +177,38 @@ function backup(){
 /*
 ** funcion para realizar backup de base de datos
 */
-function dumpMysql($conn){
+function dumpMysql($conn,$dbase){
 
     if($conn){
-    
-        shell_exec('../../dump_data_base.sh');
-         echo '<div class="alert alert-success" role="alert">';
-        echo '<h1 class="panel-title text-left" contenteditable="true"><img src="../../icons/actions/dialog-ok-apply.png"  class="img-reponsive img-rounded"><strong> Dump Successfully!!!</strong></h1>';
-         echo "</div>";
+        
+        $fecha = date('d-m-Y');
+        $salida = $dbase.'-'.$fecha.'.sql';
+        
+        $dump = "mysqldump --user='root' --password='slack142' --host='slackzone.ddns.net' $dbase > $salida";
+        system($dump);
+        
+        if(is_file($salida)){
+        
+         echo '<div class="alert alert-success" role="alert">
+               <h1 class="panel-title text-left" contenteditable="true">
+                <img src="../../icons/actions/dialog-ok-apply.png"  class="img-reponsive img-rounded"><strong> Dump Successfully!!!</strong></h1>
+                </div>';
+            copy($salida, '../../sqls/'.$salida);
+            unlink($salida);
+         
+         }else{
+            echo '<div class="alert alert-danger" role="alert">
+                    <h1 class="panel-title text-left" contenteditable="true">
+                        <img src="../../icons/actions/application-exit.png"  class="img-reponsive img-rounded"><strong> No se pudo hacer back up de la base de datos</strong></h1>
+                 </div>';
+         
+         }
         
          }else{
             
             echo '<div class="alert alert-danger" role="alert">';
-            echo '<h1 class="panel-title text-left" contenteditable="true"><img src="../../icons/actions/dialog-ok-apply.png"                   class="img-reponsive img-rounded"><strong>'. mysqli_error($conn). '</strong></h1>';
+            echo '<h1 class="panel-title text-left" contenteditable="true">
+                    <img src="../../icons/actions/dialog-ok-apply.png" class="img-reponsive img-rounded"><strong>'. mysqli_error($conn). '</strong></h1>';
          
          }
